Add new PID method
Add a new method that uses PID to drive the robot forward the specified number of inches. Unlike the existing method for this, this method does not take PID gains as arguments, and instead uses the PID values outlined in the Drivetrain constants. This will make writing autonomous code easier, as we won't have to specify the gains every time we call the method but rather can use the existing gains.