Fix PID loop auto exit
Fix a bug which caused the PID loop to be exited the first time the method hasReachedTarget() been wrong, if we hadn't yet calculated our error. This was do the previous error variable being set to - by default, and checking to see if the previousError value was within the threshold (and therefore equaled zero) before the calculate method was run resulted in this method returning true, therefore skipping the PID loop in its entirety. This commit fixes this by adding a check for the default previousTime in the hasReachedTarget method.