Fix gyro issue
Apparently the IMU in the Rev Hub returns values ranging from -180 to 180 for rotation in the y axis. These values are rotated for our robot, so all movements clockwise are negative when they should be positive, and vise versa. On top of this, instead of working between the values of 0 and 360 degrees, the REV Hub works between -180 and 180. This commit adds code for the y axis only that compensates for this and also for the fact that the values are inverted. This means that we'll always get a value between 0 and 360, and it will reset once we get passed 0 or 360.