Implement new armdex
This commit does a variety of things, but it's difficult to split those things into multiple commits. Consequently, this commit was made. This commit adds the intake color sensor back in, since it is now onboard the robot. This commit also adds more teleop documentation, adds an encoder reset to the init method of the armdex, updates the intake sensor values with the proper ones, resorts the wristUp method back to its original contents, and adds a drive power multiplier of 0.5 to help fight the power draw of our drivetrain.