LED Rewrite
Rewrite the LedController class to minimize clutter and to make the class more capable. This initial commit of the new class contains a "nextColor" method that allows you to switch through color in sequence. This could be useful for autonomous where we can switch colors after every action is performed just to show that something is happening.