Remove "init" method and replace with constructor
While FIRST does use the "init" method to pass information on to newly created objects instead of the constructor, this approach has proven to be time consuming and has not served a purpose for our team. Users don't interact with any FIRST provided code that uses the "init" method, so it's not as if we're used to including it when instantiating a new object. Furthermore, this approach results in the initialization code being longer than the autonomous code in a few of our programs. Programming autonomous runs takes time not because you're programming the actions but rather because you're intializing the mechanisms.